+++ Demo/src/main/java/com/braintreepayments/demo/PayPalFragment.java
@@ -49,6 +49,7 @@ public class PayPalFragment extends BaseFragment {
         TextInputEditText buyerPhoneNationalNumberEditText = view.findViewById(R.id.buyer_phone_national_number_edit_text);
         Button billingAgreementButton = view.findViewById(R.id.paypal_billing_agreement_button);
         Button singlePaymentButton = view.findViewById(R.id.paypal_single_payment_button);
+        Switch offerPayLater = view.findViewById(R.id.offer_pay_later_switch);
         Switch contactInformationSwitch = view.findViewById(R.id.contact_info_switch);
         Switch amountBreakdownSwitch = view.findViewById(R.id.amount_breakdown_switch);
 
@@ -59,6 +60,7 @@ public class PayPalFragment extends BaseFragment {
                 buyerPhoneCountryCodeEditText.getText().toString(),
                 buyerPhoneNationalNumberEditText.getText().toString(),
                 contactInformationSwitch.isChecked(),
+                offerPayLater.isChecked(),
                 amountBreakdownSwitch.isChecked()
             );
         });
@@ -69,6 +71,7 @@ public class PayPalFragment extends BaseFragment {
                 buyerPhoneCountryCodeEditText.getText().toString(),
                 buyerPhoneNationalNumberEditText.getText().toString(),
                 false,
+                offerPayLater.isChecked(),
                 amountBreakdownSwitch.isChecked()
             );
         });
@@ -117,6 +120,7 @@ public class PayPalFragment extends BaseFragment {
         String buyerPhoneCountryCode,
         String buyerPhoneNationalNumber,
         Boolean isContactInformationEnabled,
+        Boolean offerPayLater,
         Boolean isAmountBreakdownEnabled
     ) {
         FragmentActivity activity = getActivity();
@@ -129,10 +133,10 @@ public class PayPalFragment extends BaseFragment {
                 if (dataCollectorResult instanceof DataCollectorResult.Success) {
                     deviceData = ((DataCollectorResult.Success) dataCollectorResult).getDeviceData();
                 }
-                launchPayPal(activity, isBillingAgreement, amount, buyerEmailAddress, buyerPhoneCountryCode, buyerPhoneNationalNumber, isContactInformationEnabled, isAmountBreakdownEnabled);
+                launchPayPal(activity, isBillingAgreement, amount, buyerEmailAddress, buyerPhoneCountryCode, buyerPhoneNationalNumber, isContactInformationEnabled, offerPayLater, isAmountBreakdownEnabled);
             });
         } else {
-            launchPayPal(activity, isBillingAgreement, amount, buyerEmailAddress, buyerPhoneCountryCode, buyerPhoneNationalNumber, isContactInformationEnabled, isAmountBreakdownEnabled);
+            launchPayPal(activity, isBillingAgreement, amount, buyerEmailAddress, buyerPhoneCountryCode, buyerPhoneNationalNumber, isContactInformationEnabled, offerPayLater, isAmountBreakdownEnabled);
         }
     }
 
@@ -143,8 +147,9 @@ public class PayPalFragment extends BaseFragment {
         String buyerEmailAddress,
         String buyerPhoneCountryCode,
         String buyerPhoneNationalNumber,
-        Boolean isContactInformationEnabled,
-        Boolean isAmountBreakdownEnabled
+        boolean isContactInformationEnabled,
+        boolean offerPayLater,
+        boolean isAmountBreakdownEnabled
     ) {
         PayPalRequest payPalRequest;
         if (isBillingAgreement) {
@@ -164,6 +169,7 @@ public class PayPalFragment extends BaseFragment {
                 buyerPhoneNationalNumber,
                 isContactInformationEnabled,
                     null,
+                    offerPayLater,
                     isAmountBreakdownEnabled
             );
         }