+++ SharedUtils/src/main/java/com/braintreepayments/api/sharedutils/UnprocessableEntityException.kt
@@ -1,0 +1,7 @@
+package com.braintreepayments.api.sharedutils
+
+/**
+ * Exception thrown when a 422 HTTP_UNPROCESSABLE_ENTITY response is encountered. Indicates the
+ * request was invalid in some way.
+ */
+class UnprocessableEntityException(message: String) : Exception(message)