+++ TestUtils/src/main/java/com/braintreepayments/api/testutils/SharedPreferencesHelper.java
@@ -35,8 +35,4 @@ public class SharedPreferencesHelper {
                 .putStringAndLong(cacheKey, configuration.toJson(), timestampKey,
                         System.currentTimeMillis());
     }
-
-    public static void clearConfigurationCacheOverride(Context context) {
-        BraintreeSharedPreferences.getInstance(context).clearSharedPreferences();
-    }
 }